# gateway/openapi/vehicule-openapi.yaml
openapi: 3.0.3
info:
  title: Service Véhicules API
  description: Gestion du parc automobile — CRUD, statut, disponibilité, historique
  version: 1.0.0

servers:
  - url: http://localhost:3001/api/vehicules
    description: Développement local

security:
  - oauth2Keycloak: [utilisateur]

paths:

  /vehicules:
    get:
      summary: Liste tous les véhicules
      tags: [Véhicules]
      parameters:
        - name: statut
          in: query
          schema: { $ref: '#/components/schemas/StatutVehicule' }
        - name: marque
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Liste paginée des véhicules
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Vehicule' }
                  total: { type: integer }
                  page:  { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }

    post:
      summary: Créer un nouveau véhicule
      tags: [Véhicules]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateVehiculeDto' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vehicule' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: Immatriculation déjà existante

  /vehicules/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      summary: Récupérer un véhicule par ID
      tags: [Véhicules]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vehicule' }
        '404': { $ref: '#/components/responses/NotFound' }

    put:
      summary: Mettre à jour un véhicule
      tags: [Véhicules]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateVehiculeDto' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vehicule' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      summary: Supprimer un véhicule
      tags: [Véhicules]
      responses:
        '204': { description: Véhicule supprimé }
        '404': { $ref: '#/components/responses/NotFound' }

  /vehicules/{id}/statut:
    patch:
      summary: Changer le statut d'un véhicule
      description: Déclenche un événement Kafka fleet.vehicules.statut
      tags: [Véhicules]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [statut]
              properties:
                statut: { $ref: '#/components/schemas/StatutVehicule' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vehicule' }
        '404': { $ref: '#/components/responses/NotFound' }

  /vehicules/{id}/kilometrage:
    patch:
      summary: Mettre à jour le kilométrage
      tags: [Véhicules]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kilometrage]
              properties:
                kilometrage: { type: integer, minimum: 0 }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vehicule' }

  /vehicules/disponibles:
    get:
      summary: Liste uniquement les véhicules disponibles
      tags: [Véhicules]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Vehicule' }

components:
  securitySchemes:
    oauth2Keycloak:
      type: oauth2
      description: OAuth2 Authorization Code via Keycloak realm fleet-management
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8080/realms/fleet-management/protocol/openid-connect/auth
          tokenUrl: http://localhost:8080/realms/fleet-management/protocol/openid-connect/token
          refreshUrl: http://localhost:8080/realms/fleet-management/protocol/openid-connect/token
          scopes:
            admin: Acces administrateur complet
            manager: Gestion operationnelle de la flotte
            technicien: Gestion des interventions de maintenance
            utilisateur: Consultation et operations standard
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    StatutVehicule:
      type: string
      enum: [disponible, en_course, en_maintenance]

    Vehicule:
      type: object
      properties:
        id_vehicule:     { type: string, format: uuid }
        immatriculation: { type: string, example: "AB-123-CD" }
        marque:          { type: string, example: "Renault" }
        modele:          { type: string, example: "Clio" }
        annee:           { type: integer, example: 2022 }
        statut:          { $ref: '#/components/schemas/StatutVehicule' }
        kilometrage:     { type: integer, example: 45000 }
        created_at:      { type: string, format: date-time }

    CreateVehiculeDto:
      type: object
      required: [immatriculation, marque, modele, annee]
      properties:
        immatriculation: { type: string, example: "AB-123-CD" }
        marque:          { type: string, example: "Renault" }
        modele:          { type: string, example: "Clio" }
        annee:           { type: integer, example: 2022 }
        kilometrage:     { type: integer, default: 0 }

    UpdateVehiculeDto:
      type: object
      properties:
        marque:      { type: string }
        modele:      { type: string }
        annee:       { type: integer }
        kilometrage: { type: integer }

  responses:
    Unauthorized:
      description: Token JWT manquant ou invalide
    BadRequest:
      description: Données invalides
    NotFound:
      description: Ressource introuvable